refactor: add path aliases, re-export files and update imports
