+++ src/common/validators/index.ts
@@ -1,0 +1,1 @@
+export * from './is-military-time.validator';
