+++ src/modules/app.module.ts
@@ -5,17 +5,18 @@ import { ConfigModule } from '@nestjs/config';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 
-import { WorkingHoursModule } from './shop/working_hours/working_hours.module';
-import { ShopModule } from './shop/shop/shop.module';
-import { ShopCategoryModule } from './shop/shop-category/shop-category.module';
-import { ProductModule } from './product/product.module';
-import { CategoryModule } from './category/category.module';
+import { WorkingHoursModule } from '@shop/working_hours/working_hours.module';
+import { ShopModule } from '@shop/shop/shop.module';
+import { ShopCategoryModule } from '@shop/shop-category/shop-category.module';
+import { ProductModule } from '@product/product.module';
+import { CategoryModule } from '@category/category.module';
 import { UserModule } from '@user/user.module';
-import { AuthModule } from './auth/auth.module';
-import { JwtAuthGuard } from './auth/guards/jwt-auth.guard';
-import { APP_GUARD } from '@nestjs/core';
+import { AuthModule } from '@auth/auth.module';
 import { SendgridModule } from '@sendgrid/sendgrid.module';
 import { MailModule } from '@mail/mail.module';
+
+import { JwtAuthGuard } from '@auth/guards';
+import { APP_GUARD } from '@nestjs/core';
 
 @Module({
   imports: [
