+++ src/modules/auth/dto/index.ts
@@ -1,0 +1,3 @@
+export * from './google-auth-user.dto';
+export * from './login-user.dto';
+export * from './register-user.dto';
