+++ src/modules/auth/entities/token.entity.ts
@@ -1,4 +1,4 @@
-import { User } from '@user/entities/user.entity';
+import { User } from '@user/entities';
 import { Column, Entity, ManyToOne, PrimaryGeneratedColumn } from 'typeorm';
 
 @Entity()
