+++ src/modules/auth/guards/email-verified.guard.ts
@@ -7,7 +7,7 @@ import {
 } from '@nestjs/common';
 import { Reflector } from '@nestjs/core';
 import { Observable } from 'rxjs';
-import { isPublicRoute } from '@utils';
+import { isPublicRoute } from '@common/utils';
 
 @Injectable()
 export class EmailVerifiedGuard implements CanActivate {
