+++ src/modules/auth/guards/index.ts
@@ -1,0 +1,3 @@
+export * from './email-verified.guard';
+export * from './google-auth.guard';
+export * from './jwt-auth.guard';
