+++ src/modules/auth/strategies/jwt.strategy.ts
@@ -2,7 +2,7 @@ import { ConfigService } from '@nestjs/config';
 import { ExtractJwt, Strategy } from 'passport-jwt';
 import { PassportStrategy } from '@nestjs/passport';
 import { Injectable, UnauthorizedException } from '@nestjs/common';
-import { JWTPayload } from '../types';
+import { JWTPayload } from '@auth/types';
 import { UserService } from '@user/user.service';
 
 @Injectable()
