+++ src/modules/mail/dto/index.ts
@@ -1,0 +1,1 @@
+export * from './email-verification.dto';
